requirements.txt
streamlit==1.22.0
fastapi==0.95.2
uvicorn==0.22.0
requests==2.30.0
python-dotenv==1.0.0
groq==0.3.0
pydantic==1.10.7